move the scheduler into its own module
