+++ src/leadership/mod.rs
@@ -1,10 +1,12 @@
 mod epoch_parameters;
 pub mod leaderships;
 mod process;
+mod schedule;
 mod task;
 
 pub use self::leaderships::*;
 
 pub use self::epoch_parameters::EpochParameters;
 pub use self::process::{HandleEpochError, Process, ProcessError};
+pub use self::schedule::{LeaderSchedule, ScheduledEvent};
 pub use self::task::{Task, TaskParameters};
